+++ esa_climate_toolbox/util/im/cschemes.py
@@ -39,9 +39,10 @@ from xcube.core.gridmapping.helpers import Number
 
 _LOG = logging.getLogger('ect')
 
-LAND_COVER_CCI_CMAP = 'land_cover_cci'
-LAND_COVER_FIRE_CCI_MAP = "land_cover_fire_cci"
-HIGH_RES_LAND_COVER_CCI_CMAP = 'highres_land_cover_cci'
+
+def random_string():
+    characters = string.ascii_lowercase + string.digits
+    return ''.join(random.choices(characters, k=6))
 
 
 class ColorScheme(ABC):
@@ -56,7 +57,6 @@ class ColorScheme(ABC):
         self._labels = cmap_labels
         self._colors = cmap_colors
         self._cmap = None
-        self._register_color_map()
 
     @property
     def color_scheme_name(self) -> str:
@@ -148,6 +148,79 @@ class CategoricalColorScheme(ColorScheme):
         return (color_bounds[:-1] + color_bounds[1:]) / 2
 
 
+class CategoricalContinuousColorScheme(CategoricalColorScheme):
+    """
+    A color scheme that consists partly of categorical data and partly of continuous data.
+    This concerns data arrays that reserve a range of their values for flags and another range
+    for other data (e.g., Fire JD or Snow SWE).
+    """
+
+    def __init__(
+            self, color_scheme_name: str, cmap_name_base: str, cat_values: List[Number],
+            cat_colors: List[str], cat_labels: List[str] = None, cmap_cont_name: str = "viridis"
+    ):
+        """
+        Creates a categorical-continuous colour scheme. To the outside, it appears like a categorical scheme,
+        where parts of the range are assigned to values from a continuous range.
+        Its values, labels and colours are not fixed but adjusted dynamically by setting the values in the continuous
+        range using the method `set_continuous_values`.
+
+        :param color_scheme_name: The name of the color scheme
+        :param cmap_name_base: The base for building a color map name. As the maps are created dynamically,
+            a new map will be created and registered on every change of values.
+        :param cat_values: The values of the categorical part of the color scheme.
+        :param cat_colors: The colors of the categorical part of the color scheme.
+        :param cat_labels: The labels of the categorical part of the color scheme, may  be none
+        :param cmap_cont_name: The name of the color map which shall be used to encode the continuous part
+            of the values. Default is `viridis`.
+        """
+        self._cmap_name_base = cmap_name_base
+        self._base_values = cat_values
+        self._base_colors = cat_colors
+        self._base_labels = cat_labels
+        self._cmap_cont_name = cmap_cont_name
+        super().__init__(
+            color_scheme_name, f"{cmap_name_base}_base", cat_values, cat_colors, cat_labels
+        )
+
+    def set_continuous_values(
+            self, cont_values: List[Number], range_extent: int = None, label_shift: int = 0, labels_space: int = 5
+    ):
+        """
+        Lists the values to be encoded by values from the continuous color map.
+
+        :param cont_values: The list of values
+        :param range_extent: Optional extent of the range to apply to the color map.
+            Set to the same value to make values from different time steps comparable,
+            otherwise the color map is set to cover exactly the value range.
+        :param label_shift: Shifts the labels of the continuous values from their actual
+            value by subtracting the passed shift.
+            Useful for, e.g., displaying days of year to days of months.
+        :param labels_space: Indicates how often a value in the continuous part of the
+            range shall be labeled. Default is 5.
+        """
+        self._values = np.array(self._base_values + cont_values)
+
+        num_values = len(cont_values)
+        range_extent = range_extent or num_values
+        cont_cmap = matplotlib.colormaps[self._cmap_cont_name]
+        cont_colors = cont_cmap(np.linspace(0, 1, num=range_extent))
+        intermediate_colors = self._base_colors + cont_colors.tolist()
+        self._colors = intermediate_colors[:len(self._base_values) + num_values]
+
+        cont_labels = [""] * len(cont_values)
+        for i in range(labels_space - 1, num_values, labels_space):
+            cont_labels[i] = str(cont_values[i] - label_shift)
+        self._labels = self._base_labels + cont_labels
+
+        self._cmap = None
+        color_map_name = f"{self._cmap_name_base}_{random_string()}"
+        while color_map_name in matplotlib.colormaps:
+            color_map_name = f"{self._cmap_name_base}_{random_string()}"
+        self._color_map_name = color_map_name
+        self._register_color_map()
+
+
 class ColorSchemeRegistry:
 
     def __init__(self):
@@ -166,15 +239,12 @@ class ColorSchemeRegistry:
         )
 
     @staticmethod
-    def random_part():
-        characters = string.ascii_lowercase + string.digits
-        return ''.join(random.choices(characters, k=6))
+    def random_cmap_name():
+        return f"cmap_{random_string()}"
 
-    def random_cmap_name(self):
-        return f"cmap_{self.random_part()}"
-
-    def random_cscheme_name(self):
-        return f"cscheme_{self.random_part()}"
+    @staticmethod
+    def random_cscheme_name():
+        return f"cscheme_{random_string()}"
 
     def register_categorical_color_scheme(
             self, cm_values: List, color_scheme_name: str = None, color_map_name: str = None,
@@ -198,6 +268,32 @@ class ColorSchemeRegistry:
                 cm_entry["label"] = cm_labels[i]
             cm_entries.append(cm_entry)
         self._register_categorical_color_scheme(color_scheme_name, color_map_name, cm_entries)
+        return color_scheme_name
+
+    def register_categorical_continuous_color_scheme(
+            self, cm_values: List, color_scheme_name: str = None, color_map_name_base: str = None,
+            cont_color_map_name: str = "viridis", cm_colors: List [str] = None, cm_labels: List[str] = None
+    ) -> str:
+        if color_scheme_name is None:
+            color_scheme_name = self.random_cscheme_name()
+            while self.has_color_scheme(color_scheme_name):
+                color_scheme_name = self.random_cscheme_name()
+        if color_map_name_base is None:
+            color_map_name = self.random_cmap_name()
+            while color_map_name in matplotlib.colormaps:
+                color_map_name = self.random_cmap_name()
+        cm_entries = []
+        for i, cm_value in enumerate(cm_values):
+            cm_entry = dict()
+            cm_entry["value"] = cm_value
+            if cm_colors:
+                cm_entry["color"] = cm_colors[i]
+            if cm_labels:
+                cm_entry["label"] = cm_labels[i]
+            cm_entries.append(cm_entry)
+        self._register_categorical_continuous_color_scheme(
+            color_scheme_name, color_map_name_base, cont_color_map_name, cm_entries
+        )
         return color_scheme_name
 
     def deregister_color_scheme(self, cs_name: str):
@@ -228,6 +324,25 @@ class ColorSchemeRegistry:
         ccs = CategoricalColorScheme(color_scheme_name, color_map_name, values, colors, labels)
         self._colorschemes[color_scheme_name] = ccs
 
+    def _register_categorical_continuous_color_scheme(
+            self, color_scheme_name: str, color_map_base_name: str,
+            color_map_continuous_name: str, cm_entries: List[Dict]
+    ):
+        values = []
+        labels = []
+        colors = []
+        for cm_entry in cm_entries:
+            values.append(cm_entry.get("value"))
+            labels.append(cm_entry.get("label", ""))
+            if "color" in cm_entry:
+                colors.append(cm_entry.get("color"))
+            else:
+                colors.append(self._random_color())
+        cccs = CategoricalContinuousColorScheme(
+            color_scheme_name, color_map_base_name, values, colors, labels, color_map_continuous_name
+        )
+        self._colorschemes[color_scheme_name] = cccs
+
 COLOR_SCHEME_REGISTRY = ColorSchemeRegistry()
 
 def ensure_cschemes_loaded():
@@ -240,5 +355,17 @@ def ensure_cschemes_loaded():
             cm_name = cs_values["color_map_name"]
             cs_entries = cs_values["entries"]
             COLOR_SCHEME_REGISTRY._register_categorical_color_scheme(cs_name, cm_name, cs_entries)
+    categorical_continuous_colorschemes_file = os.path.join(dir_path, 'data/categorical_continuous_colorschemes.yml')
+    with open(categorical_continuous_colorschemes_file, "r") as cs:
+        color_schemes_dict = yaml.safe_load(cs)
+    for cs_name, cs_values in color_schemes_dict.items():
+        if not COLOR_SCHEME_REGISTRY.has_color_scheme(cs_name):
+            cm_base_name = cs_values["cmap_name_base"]
+            cm_cont_name = cs_values["cmap_cont_name"]
+            cs_entries = cs_values["entries"]
+            COLOR_SCHEME_REGISTRY._register_categorical_continuous_color_scheme(
+                cs_name, cm_base_name, cm_cont_name, cs_entries
+            )
+
 
 ensure_cschemes_loaded()